Upgrade rust edition and cargo, git2 dependencies and fix for API changes
It was necessary to upgrade cargo when upgrading rust edition. After that
git2 required upgrade as well to match the version used by cargo (see
detailed error message below).

Some fixes were made to match API changes in the cargo lib:

* `source.update()` is now `source.invalidate_cache()`.
* We now must specify fuzzy or exact when calling `source.query_vec(...)`.
  Set to `QueryKind::Exact` because that was the default in effect before.
* Now matching on result of `source.query_vec(...)`, and blocking until
  ready if query_vec result is `Pending`. It is also necessary to acquire
  a package lock for this to pass tests.
* Now displaying error message when error message assertions fail in
  `wrong_version` test, so the original backtrace is visible on failure.

Here is the error seen when cargo was upgraded on its own without also
upgrading git2:

```
the package `libgit2-sys` links to the native library `git2`, but it conflicts with a previous package which links to `git2` as well:
package `libgit2-sys v0.14.1+1.5.0`
    ... which satisfies dependency `libgit2-sys = "^0.14.1"` of package `cargo v0.68.0`
    ... which satisfies dependency `cargo = "^0.68"` of package `cargo-edit-locally v0.2.0 (/Users/melo/Src/cargo-edit-locally)`
```